fix: FORWARD_DAYS_SCHEDULE in plans. Если поле результата пусто не добавлять
